refactor(daemon): extract deferred turn-finalize into its own module
The deferred end-of-turn work added in the previous commit landed inline in
`conversation-agent-loop.ts` (1.8k lines) and `conversation-agent-loop-handlers.ts`
(2.7k lines), adding to files that are already too big. Pull the cohesive unit —
the per-message finalize side-effect builder and the end-of-turn deferred tail —
into a focused `conversation-turn-finalize.ts` that both callers import.

Behavior is unchanged: `handleMessageComplete` now calls
`buildDeferredFinalizeEffect(...)` to construct the closure it queues, and the
orchestrator calls `runDeferredTurnTail(...)`. Net -174 lines across the two
large files. Drops the now-unused truncation/disk-view/attention/sync imports
from the callers. No new circular dependencies (module is in zero cycles;
`lint:circular` count unchanged at 402).

Part of LUM-2654.
